+++ R/resampledata.R
@@ -1,7 +1,10 @@
 #' resampledata: Data Sets for Mathematical Statistics with Resampling in R
 #'
 #' Package of data sets from "Mathematical Statistics with Resampling in R"
-#' (2011) by Laura Chihara and Tim Hesterberg.
+#' (2011) by Laura Chihara and Tim Hesterberg. Note the \code{\link{Bangladesh}}
+#' data is provided solely for illustrative purposes and to enable statistical
+#' analysis. Full data are available from the British Geological Survey web site
+#' \url{http://www.bgs.ac.uk/arsenic/bphase2/datadownload.htm}.
 #'
 #' @section Data sets:
 #' Type \code{data(package="resampledata")} to view all datasets.
@@ -10,6 +13,16 @@
 #' @examples
 #' library(resampledata)
 #' data(FlightDelays)
+#' @references Hesterberg, Tim (2014), What Teachers Should Know about the
+#' Bootstrap: Resampling in the Undergraduate Statistics Curriculum,
+#' \url{http://arxiv.org/abs/1411.5279}.
+#' @references Chihara, Laura M., Hesterberg, Tim C. (2011). Mathematical
+#'   Statistics with Resampling and R. Hoboken, NJ: John Wiley & Sons.
 #' @name resampledata
 NULL
 #> NULL
+
+
+
+
+